backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
qrcode[pil]
user-agents
geoip2
requests
strawberry-graphql
